Fixed GA by making predictions start at random vehicle cutoff
* GA
Created a local variable data, which gets the previous vehicle data and then appends the predictions
to it, creating a local set of data for every individual's fitness prediction. This solved the problem of
the fitness calculation always increasing in time.

Removed the calculation of fitness for both individuals in tournament(), instead only gets fitness
of loser after crossover. This makes the number of get_fitness() calls per iterations down to 1.

Only changed the fitness of the loser in the pool

Made method return best individual's genome

Removed variable all_fitness and renamed it pool

Put sprite creation from get_fitness to the if condition

* Cycle
Removed seed from testing, which made random movements for our vehicle not random