Create outline of sleep_wake function. The vehicle will iterate through the two states.
*Cycle
The method uses the wake learning function for the initial movement.
It handles the data returned from the GA and the ones returned from the simulator.
Keeps the sprites updated.
Need to handle the changes in the fittest brain.
Introduce noise to the brain movement for the production of training data for the next model.