Run a multi-vehicle single trajectory to train the model and  predict from the last position. Generalize GA fitness with vehicles fro random positions.
*Cycle
Declare variables for the the position, angle and past data of the last vehicle trajectory.
Save them after collecting the data for the training and fitness eval.
Sleep_wake set the last position, angle and past data to be those of the last vehicle trajectory.

*Genetic
Make the initialization of the genomes to be in the range or half the max genome scale.